% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_construction.R
\name{icr_overlap}
\alias{icr_overlap}
\title{ICR Overlap}
\usage{
icr_overlap(x, sample = NULL, percent = NULL, coders)
}
\arguments{
\item{x}{A dataframe containing the data.}

\item{sample}{A whole number specifying how many observations to overlap between coders.}

\item{percent}{A percentage expressed as a decimal between 0 and 1 that specifies the percentage of observations to sample.
This argument cannot be used with the sample argument.}

\item{coders}{A whole number specifying how many coders to assign observations to.}
}
\value{
A modified dataframe with a new column "coder" indicating coder assignments.
}
\description{
This function creates a dataset assigning coders to observations with a given level of overlap between coders.
It randomly selects a given number of rows, and assigns them to all coders.
It attempts to evenly distribute the remaining observations across all coders.
}
\examples{
df <- data.frame(a = 1:10, b = letters[1:10])
icr_overlap(df, sample = 5, coders = 3)
}
