% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_construction.R
\name{icr_random_draw}
\alias{icr_random_draw}
\title{ICR Random Draw}
\usage{
icr_random_draw(x, sample)
}
\arguments{
\item{x}{A dataframe containing the data.}

\item{sample}{A whole number specifying how many rows to randomly sample.}
}
\value{
A modified dataframe with randomly selected rows.
}
\description{
This function randomly selects observations from a dataframe to subset a given number for ICR analysis.
}
\examples{
df <- data.frame(a = 1:10, b = letters[1:10])
icr_random_draw(df, sample = 5)
}
