% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr_sampling.R
\name{lacy_riffe_sample}
\alias{lacy_riffe_sample}
\title{Lacy Riffe Sample}
\usage{
lacy_riffe_sample(x, ci = 0.95, min.agree = 0.85)
}
\arguments{
\item{x}{A dataframe from which to sample observations.}

\item{ci}{A numeric value representing the confidence interval (default is 0.95).}

\item{min.agree}{A numeric value representing the minimum acceptable agreement (default is 0.85).}
}
\value{
A randomly selected subset of observations from the dataframe.
}
\description{
This function uses the `lacy_riffe_calc` function to subset a random selection of observations
from a dataframe. The number of observations selected is based on the result from `lacy_riffe_calc()`.
}
\examples{
df <- data.frame(a = 1:100, b = rnorm(100))
sampled_data <- lacy_riffe_sample(df)
}
